@@ -1,6 +1,7 @@
 package com.tamaproject.util;
 
 
+import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
@@ -8,13 +9,17 @@ import java.net.URL;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import android.annotation.SuppressLint;
+import android.util.Log;
+
+import javax.net.ssl.HttpsURLConnection;
 
 public class handlejson {
 
-   private String weather = "none";
-   private String main = "none";
-   private String temp = "temperature";
-
+   private static final String TAG = "Tama Game handlejson";
+   private String agent = "my own written app email: elymbmx@gmail.com";
+   private String temp = "Temp";
+   private String weather = "Weather";
+   //private String icon = "Weatherimage";
    private String urlString = null;
 
    public volatile boolean parsingComplete = true;
@@ -23,66 +28,71 @@ public class handlejson {
    }
 
 
-   public String getWeather() {
-      return weather;
-   }
-   public String getmain() {
-        return main;
-    }
    public String getTemp() {
       return temp;
    }
+   public String getWeather() {
+      return weather;
+   }
+   //public String getIcon() { return icon; }
 
    @SuppressLint("NewApi")
    public void readAndParseJSON(String in) {
       try {
          JSONObject reader = new JSONObject(in);
 
-         //JSONObject getweather  = reader.getJSONObject("weather");
-         //get weather
-         JSONArray weatherarray = reader.getJSONArray("weather");
-         JSONObject weatherobject = weatherarray.getJSONObject(0);
-         main = weatherobject.getString("main");
-         weather = weatherobject.getString("description");
-         //get temp
-         JSONObject gettemp  = reader.getJSONObject("main");
-         temp = gettemp.getString("temp");
+         JSONObject current  = reader.getJSONObject("currentobservation");
+         temp = current.getString("Temp");
+         weather = current.getString("Weather");
+         //icon = current.getString("Weatherimage");
          parsingComplete = false;
 
-        } catch (Exception e) {
-           // TODO Auto-generated catch block
-           e.printStackTrace();
-        }
+      } catch (Exception e) {
+         Log.i(TAG, "failed to readAndParseJSON(...)...");
+         e.printStackTrace();
+      }
 
    }
    public void fetchJSON(){
       Thread thread = new Thread(new Runnable(){
          @Override
          public void run() {
-         try {
-            URL url = new URL(urlString);
-            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
-            conn.setReadTimeout(15000 /* milliseconds */);
-            conn.setConnectTimeout(15000 /* milliseconds */);
-            conn.setRequestMethod("GET");
-            conn.setDoInput(true);
-            // Starts the query
-            conn.connect();
-         InputStream stream = conn.getInputStream();
+            try {
+               URL url = new URL(urlString);
+               HttpsURLConnection conn = (HttpsURLConnection) url.openConnection();
+               Log.d(TAG, agent);
+               conn.setRequestProperty("User-Agent", agent);
+               conn.setRequestProperty("Accept", "application/vnd.noaa.dwml+xml;version=1");
 
-      String data = convertStreamToString(stream);
+               conn.setReadTimeout(10000 /* milliseconds */);
+               conn.setConnectTimeout(10000 /* milliseconds */);
+               conn.setRequestMethod("GET");
+               conn.setDoInput(true);
+               // Starts the query
+               conn.connect();
+               InputStream stream = conn.getInputStream();
 
-      readAndParseJSON(data);
-         stream.close();
+               String data = convertStreamToString(stream);
 
-         } catch (Exception e) {
-            e.printStackTrace();
-         }
+               readAndParseJSON(data);
+               stream.close();
+            } catch (NullPointerException np) {
+               Log.i(TAG, "NullPointerException in fetchJSON()...");
+               np.printStackTrace();
+            } catch (IOException io) {
+               Log.i(TAG, "IOException in fetchJSON()...");
+               io.printStackTrace();
+            } catch (Exception e) {
+               Log.i(TAG, "Exception in fetchJSON()...");
+               e.printStackTrace();
+            }
          }
       });
-
-       thread.start(); 		
+      Log.i(TAG, "fetchJSON() thread start");
+      thread.start();
+      Log.i(TAG, "fetchJSON() end...");
    }
+
    static String convertStreamToString(java.io.InputStream is) {
       java.util.Scanner s = new java.util.Scanner(is).useDelimiter("\\A");
       return s.hasNext() ? s.next() : "";
